Completed Project 006 Animated Scrolling
Completed project 6 which features scrolling animation. This one was cool in terms of javascript. The css was nothing special, but the javascript had some things id never seen before. Particularly the window scroll event listener, which allows you to call an event whenever the user scrolls on the page.. Also the window inner height property, which checks where the window is, and the getBoundingClientRect() function which tells you where an element is on a page.