Completed Project 007 Split Landing Page
Added in the basic HTML and CSS of the split landing page. Got some more experience working with variables in CSS and using rbga values. Also got to learn some new unique CSS properties, such as the Text-Transform: uppercase property which makes everything become uppercase. Or the whitespace nowrap property. Also added in Javascript that utilize the mouseenter and mouseleave events so that when the user hovers over each side the transition occurs. Also used a media query for when dealing with small screens.