Completed Project 008 Form Wave Animations
Created the log in form. Getting better with styling, and adding my own tweaks here and there. Also the experimentation is giving me more experience with how the styles work together and how to troubleshoot when things don’t look as I desire. Learned about the split, map, and join javascript functions which allow you to split a string up into an array, map the values to new values, and join an array into a string respectively. Also used a cubic Bezier for the first time, which allows you to do transitions with a modified time curve. Subtle, but adds a sexy effect to transitions.