Completed Project 011 Event Keycodes
Did the javascript for the project all on my own!!! The keycode project. Used the event object, e, to figure out which key was pressed, then passed those values into the HTML to dynamically change the keycode information displayed. Can use this project to get keycodes for things too, so has some uses!