Completed Project 013 Choice Master 9000
Went above and beyond again! Added a ton of CSS animation into this project, and really made it my own. This one used a lot of intervals in the javascript, used for timing of the flickering choices. Also used the math.random() and math.floor() functions for the first time to calculate the random value.