Completed Project 015 Incrementing Counter
Got some more experience using icons with font awesome. In the incrementing counter project we used social media icons. Also used the data-target attribute for the first time. Id never seen it before. Learned about a new interesting way of creating an arrow function. Creating a function to be placed directly into a constant. Also learned that putting a plus sign beside a string is a quick way to parse it into an int.  Also used the setTimeout() function, which is similar to the setInterval() function, but it only runs once. This can be done in a for loop however to feature repeated runs as we did here.