Completed Project 017 Movie App
Did the CSS and styling for the movie app project. Ran into some issues, and wasn’t feeling too inspired with it, but I SHOWED UP! Completed the Movie app project. This one was really cool! Definitely going to come back to this one to add some functionality in the future. Maybe add some stuff to the header, and expand on the idea. Search by category perhaps? Can certainly make this more elaborate.